* Added: File access management feature. v1.7.2_2 Added: "ConsoleOnly" option. Fixed: Can't connect when server reload with client connected. Fixed: Some strings don't be wrapped. Fixed: Some bugs.

+++ RemoteController/src/hide92795/bukkit/plugin/remotecontroller/command/CommandFileDelete.java
@@ -3,6 +3,7 @@ package hide92795.bukkit.plugin.remotecontroller.command;
 import hide92795.bukkit.plugin.remotecontroller.ClientConnection;
 import hide92795.bukkit.plugin.remotecontroller.RemoteController;
 import hide92795.bukkit.plugin.remotecontroller.org.apache.commons.io.FileUtils;
+import hide92795.bukkit.plugin.remotecontroller.util.Util;
 import java.io.File;
 
 public class CommandFileDelete implements Command {
@@ -11,19 +12,23 @@ public class CommandFileDelete implements Command {
 		try {
 			if (connection.isAuthorized()) {
 				File file = new File(plugin.getRoot(), arg.substring(1).replace('/', File.separatorChar));
-				boolean success;
-				if (file.isDirectory()) {
-					// Directory
-					success = FileUtils.deleteQuietly(file);
+				if (Util.canWrite(plugin.config.file_access, file)) {
+					boolean success;
+					if (file.isDirectory()) {
+						// Directory
+						success = FileUtils.deleteQuietly(file);
+					} else {
+						// File
+						success = file.delete();
+					}
+					if (success) {
+						plugin.getLogger().info(connection.getUser() + " has deleted directory/file:" + arg);
+						connection.send("SUCCESS", pid, "");
+					} else {
+						connection.send("ERROR", pid, "FAILED");
+					}
 				} else {
-					// File
-					success = file.delete();
-				}
-				if (success) {
-					plugin.getLogger().info(connection.getUser() + " has deleted directory/file:" + arg);
-					connection.send("SUCCESS", pid, "");
-				} else {
-					connection.send("ERROR", pid, "FAILED");
+					connection.send("ERROR", pid, "ACCESS_DENIED");
 				}
 			} else {
 				connection.send("ERROR", pid, "NOT_AUTH");
